Reorganize a little more
