--- javaparser-core/src/main/java/com/github/javaparser/TokenBase.java
@@ -1,8 +1,0 @@
-package com.github.javaparser;
-
-import static com.github.javaparser.GeneratedJavaParserConstants.GT;
-
-class TokenBase {
-    int realKind = GT;
-    JavaToken javaToken;
-}
