+++ javaparser-core/src/main/javacc-support/com/github/javaparser/GeneratedJavaParserBase.java
@@ -5,7 +5,7 @@ import com.github.javaparser.ast.Modifier;
 import com.github.javaparser.ast.Node;
 import com.github.javaparser.ast.NodeList;
 import com.github.javaparser.ast.body.Parameter;
-import com.github.javaparser.ast.comments.*;
+import com.github.javaparser.ast.comments.CommentsCollection;
 import com.github.javaparser.ast.expr.*;
 import com.github.javaparser.ast.stmt.Statement;
 import com.github.javaparser.ast.type.ArrayType;
@@ -15,8 +15,7 @@ import com.github.javaparser.utils.Pair;
 
 import java.util.*;
 
-import static com.github.javaparser.GeneratedJavaParserConstants.*;
-import static com.github.javaparser.Position.pos;
+import static com.github.javaparser.GeneratedJavaParserConstants.EOF;
 import static com.github.javaparser.ast.type.ArrayType.unwrapArrayTypes;
 import static com.github.javaparser.ast.type.ArrayType.wrapInArrayTypes;
 import static com.github.javaparser.utils.Utils.assertNotNull;
@@ -364,35 +363,5 @@ abstract class GeneratedJavaParserBase {
                     .append(expected.toString());
         }
         return sb.toString();
-    }
-
-    ///// These are for the token manager that can't get a superclass until javacc 7.
-
-    /**
-     * Create a TokenRange that spans exactly one token
-     */
-    private static TokenRange tokenRange(Token token) {
-        JavaToken javaToken = token.javaToken;
-        return new TokenRange(javaToken, javaToken);
-    }
-
-    static Comment createCommentFromToken(Token token) {
-        String commentText = token.image;
-        if (token.kind == JAVADOC_COMMENT) {
-            return new JavadocComment(tokenRange(token), commentText.substring(3, commentText.length() - 2));
-        } else if (token.kind == MULTI_LINE_COMMENT) {
-            return new BlockComment(tokenRange(token), commentText.substring(2, commentText.length() - 2));
-        } else if (token.kind == SINGLE_LINE_COMMENT) {
-            // line comments have their end of line character(s) included, and we don't want that.
-            Range range = new Range(pos(token.beginLine, token.beginColumn), pos(token.endLine, token.endColumn));
-            while (commentText.endsWith("\r") || commentText.endsWith("\n")) {
-                commentText = commentText.substring(0, commentText.length() - 1);
-            }
-            range = range.withEnd(pos(range.begin.line, range.begin.column + commentText.length()));
-            LineComment comment = new LineComment(tokenRange(token), commentText.substring(2));
-            comment.setRange(range);
-            return comment;
-        }
-        throw new AssertionError("Unexpectedly got passed a non-comment token.");
     }
 }
