+++ javaparser-core/src/main/javacc-support/com/github/javaparser/TokenBase.java
@@ -1,0 +1,11 @@
+package com.github.javaparser;
+
+import static com.github.javaparser.GeneratedJavaParserConstants.GT;
+
+/** Base class for the generated {@link Token} */
+class TokenBase {
+    /** For tracking the >> >>> ambiguity. */
+    int realKind = GT;
+    /** This is the link to the token that JavaParser presents to the user */
+    JavaToken javaToken;
+}
